fix: move prompt path helpers out of test-only scope
Production prompt_drift.go called loadPackagedGoalRolePrompt and packagedGoalMaterializedPromptPath that lived only in _test.go, breaking make test for cmd/factory and other importers. Promote both helpers to prompt_paths.go so the drift gate compiles in non-test builds.
