+++ pkg/packagedfactories/goal/prompt_regression_test.go
@@ -462,22 +462,6 @@ func assertPackagedGoalLoadedWorkerPrompt(t *testing.T, loaded *factoryconfig.Lo
 	}
 }
 
-func packagedGoalMaterializedPromptPath(factoryDir string, source PackagedGoalRolePromptSource) string {
-	if source.SourceKind == PackagedGoalRolePromptSourceKindWorkerBody {
-		return filepath.Join(factoryDir, interfaces.WorkersDir, source.WorkerName, interfaces.FactoryAgentsFileName)
-	}
-	return filepath.Join(factoryDir, interfaces.WorkstationsDir, source.WorkstationName, source.PromptFile)
-}
-
-func loadPackagedGoalRolePrompt(factoryDir string, source PackagedGoalRolePromptSource) (string, error) {
-	promptPath := packagedGoalMaterializedPromptPath(factoryDir, source)
-	data, err := os.ReadFile(promptPath)
-	if err != nil {
-		return "", err
-	}
-	return strings.TrimSpace(string(data)), nil
-}
-
 func loadedPackagedGoalRolePrompt(loaded *factoryconfig.LoadedFactoryConfig, source PackagedGoalRolePromptSource) (string, error) {
 	switch source.SourceKind {
 	case PackagedGoalRolePromptSourceKindWorkerBody:
